# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VRF Monitor LG - API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /systems/{systemId}/startup-reports:
    get:
      operationId: listStartupReports
      tags: [startup-reports]
      summary: List startup reports for a system
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of startup reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StartupReport"
    post:
      operationId: uploadStartupReport
      tags: [startup-reports]
      summary: Upload and process a startup report (PDF, JPG, PNG or WebP)
      description: Accepts PDF or image files up to 50 MB. The AI extracts baseline data from text and/or rendered pages via GPT-4o Vision.
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/StartupReportUpload"
      responses:
        "201":
          description: Uploaded and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupReport"
  /systems/{systemId}/startup-reports/{reportId}:
    get:
      operationId: getStartupReport
      tags: [startup-reports]
      summary: Get a startup report
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Startup report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupReport"
        "404":
          description: Not found
    delete:
      operationId: deleteStartupReport
      tags: [startup-reports]
      summary: Delete a startup report
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /systems/{systemId}/startup-reports/{reportId}/reprocess:
    post:
      operationId: reprocessStartupReport
      tags: [startup-reports]
      summary: Re-run AI processing on an existing startup report
      parameters:
        - name: systemId
          in: path
          required: true
          schema:
            type: integer
        - name: reportId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reprocessing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupReport"
        "404":
          description: Not found
...
components:
  schemas:
...
    StartupReport:
      type: object
      required: [id, systemId, filename, uploadedAt, processingStatus]
      properties:
        id:
          type: integer
        systemId:
          type: integer
        filename:
          type: string
        fileUrl:
          type: ["string", "null"]
        uploadedAt:
          type: string
        processingStatus:
          type: string
          enum: [pending, processing, done, error]
        extractedData:
          type: ["object", "null"]
        errorMessage:
          type: ["string", "null"]
    StartupReportUpload:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...