# lib/api-zod/src/generated/types/startupReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VRF Monitor LG - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StartupReportExtractedData } from "./startupReportExtractedData";
import type { StartupReportProcessingStatus } from "./startupReportProcessingStatus";

export interface StartupReport {
  id: number;
  systemId: number;
  filename: string;
  /** @nullable */
  fileUrl?: string | null;
  uploadedAt: string;
  processingStatus: StartupReportProcessingStatus;
  /** @nullable */
  extractedData?: StartupReportExtractedData;
  /** @nullable */
  errorMessage?: string | null;
}